requirements.txt
python-telegram-bot==20.6
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
